Fix WASM test loading and mix-sample parity.
Inject file:// wasm bytes for Vitest, disable import.meta URL glue Vite rewrites, and keep OfflineAudioContext when media-engine is unavailable.

+++ native/media_engine/include/resampler.h
@@ -14,7 +14,7 @@ void resample_linear_stereo(
     const float* interleaved,
     int frames,
     int channels,
-    float src_frame,
+    double src_frame,
     float* out_l,
     float* out_r);
